fix replies count

+++ app/Http/Controllers/ReplyController.php
@@ -107,7 +107,7 @@ class ReplyController extends Controller
         //     return response([], 403);
         // }
         
-        //  $this->authorize('update', $reply);
+         $this->authorize('update', $reply);
 
         $reply->delete();
